refactor: reorganize agent-based simulation architecture
This commit introduces a major architectural reorganization to improve code maintainability,
reduce circular dependencies, and better separate concerns. Key changes include:

Core Changes:
- Split monolithic agents.py into separate modules
- Created new base_agent.py for core agent functionality
- Moved simulation logic to dedicated simulation.py module
- Updated imports across codebase to reflect new structure

Specific Changes:
1. Base Agent:
- Moved common agent functionality to BaseAgent class
- Centralized DQN implementation and learning mechanics
- Standardized agent initialization and resource management

2. Agent Types:
- Simplified SystemAgent and IndividualAgent to only contain type-specific behavior
- Removed duplicate code by inheriting from BaseAgent
- Maintained distinct gathering and resource sharing strategies

3. Simulation:
- Created dedicated simulation.py for simulation loop and environment management
- Improved logging and visualization setup
- Centralized configuration handling

4. Batch Runner:
- Updated to use new simulation module
- Improved parameter variation handling
- Enhanced results collection and storage
- Added better error handling and logging

5. Main Application:
- Updated imports to use new module structure
- Fixed circular dependency issues
- Maintained GUI functionality with new architecture

Performance Impact:
- Note: Current implementation may be running slower than previous version
- Performance optimization needed in future commits
- Main bottleneck likely in agent movement and learning calculations

Next Steps:
- Optimize agent movement and learning algorithms
- Add batch processing for agent updates
- Consider vectorizing common operations
- Profile and identify specific performance bottlenecks

Breaking Changes:
- Removed direct access to agent internals
- Changed module import structure
- Modified batch runner interface

Migration:
Existing code using the old structure will need to:
1. Update imports to use new modules
2. Access agent functionality through BaseAgent interface
3. Use simulation.py for running simulations
